Business/Apache/server/requirements.txt
thrift
mysql-connector-python